* big changes to plug leaks and fix initializers
rename some but not yet all underscore methods to mulle methods
* improved propertylist parsing and printing
* improved -description for containers
* adapted to changes in MulleObjC especially in regard to placeholders, singletons and deinitialization

+++ src/PropertyList/Printing/NSArray+PropertyListPrinting.m
@@ -52,7 +52,7 @@ static char   closer[]    = { ' ', ')' };
 
 
 - (void) propertyListUTF8DataToStream:(id <_MulleObjCOutputDataStream>) handle
-                               indent:(unsigned int) indent
+                               indent:(NSUInteger) indent
 {
    NSUInteger   i, n;
    id           value;
